Module datafusion::execution

This module contains the shared state available at different parts of query planning and execution

Runtime Environment

runtime_env::RuntimeEnv can be created from a runtime_env::RuntimeConfig and stores state to be shared across multiple sessions. In most applications there will be a single runtime_env::RuntimeEnv for the entire process

Session Context

context::SessionContext can be created from a context::SessionConfig and an optional runtime_env::RuntimeConfig, and stores the state for a particular query session.

In particular context::SessionState is the information available to query planning

Task Context

context::TaskContext is typically created from a context::SessionContext or context::SessionState, and represents the state available to query execution.

In particular it is the state passed to crate::physical_plan::ExecutionPlan::execute

Re-exports

pub use disk_manager::DiskManager;

pub use registry::FunctionRegistry;

Modules

context

SessionContext contains methods for registering data sources and executing queries

disk_manager

Manages files generated during query execution, files are hashed among the directories listed in RuntimeConfig::local_dirs.

memory_pool

Manages all available memory during query execution

options

User facing options for the file formats readers

registry

FunctionRegistry trait

runtime_env

Execution runtime environment that holds object Store, memory manager, disk manager and various system level components that are used during physical plan execution.