Struct datafusion::execution::context::SessionConfig

pub struct SessionConfig { /* private fields */ }

Configuration options for session context

Implementations

impl SessionConfig

pub fn new() -> Self

Create an execution config with default setting

pub fn from_env() -> Result<Self>

Create an execution config with config options read from the environment

pub fn set(self, key: &str, value: ScalarValue) -> Self

Set a configuration option

pub fn set_bool(self, key: &str, value: bool) -> Self

Set a boolean configuration option

pub fn set_u64(self, key: &str, value: u64) -> Self

Set a generic u64 configuration option

pub fn set_usize(self, key: &str, value: usize) -> Self

Set a generic usize configuration option

pub fn set_str(self, key: &str, value: &str) -> Self

Set a generic str configuration option

pub fn with_batch_size(self, n: usize) -> Self

Customize batch size

pub fn with_target_partitions(self, n: usize) -> Self

Customize target_partitions

pub fn target_partitions(&self) -> usize

Get target_partitions

pub fn information_schema(&self) -> bool

Is the information schema enabled?

pub fn create_default_catalog_and_schema(&self) -> bool

Should the context create the default catalog and schema?

pub fn repartition_joins(&self) -> bool

Are joins repartitioned during execution?

pub fn repartition_aggregations(&self) -> bool

Are aggregates repartitioned during execution?

pub fn repartition_window_functions(&self) -> bool

Are window functions repartitioned during execution?

pub fn repartition_sorts(&self) -> bool

Do we execute sorts in a per-partition fashion and merge afterwards, or do we coalesce partitions first and sort globally?

pub fn collect_statistics(&self) -> bool

Are statistics collected during execution?

pub fn with_default_catalog_and_schema( self, catalog: impl Into<String>, schema: impl Into<String> ) -> Self

Selects a name for the default catalog and schema

pub fn with_create_default_catalog_and_schema(self, create: bool) -> Self

Controls whether the default catalog and schema will be automatically created

pub fn with_information_schema(self, enabled: bool) -> Self

Enables or disables the inclusion of information_schema virtual tables

pub fn with_repartition_joins(self, enabled: bool) -> Self

Enables or disables the use of repartitioning for joins to improve parallelism

pub fn with_repartition_aggregations(self, enabled: bool) -> Self

Enables or disables the use of repartitioning for aggregations to improve parallelism

pub fn with_repartition_file_min_size(self, size: usize) -> Self

Sets minimum file range size for repartitioning scans

pub fn with_repartition_file_scans(self, enabled: bool) -> Self

Enables or disables the use of repartitioning for file scans

pub fn with_repartition_windows(self, enabled: bool) -> Self

Enables or disables the use of repartitioning for window functions to improve parallelism

pub fn with_repartition_sorts(self, enabled: bool) -> Self

Enables or disables the use of per-partition sorting to improve parallelism

pub fn with_parquet_pruning(self, enabled: bool) -> Self

Enables or disables the use of pruning predicate for parquet readers to skip row groups

pub fn parquet_pruning(&self) -> bool

Returns true if pruning predicate should be used to skip parquet row groups

pub fn with_collect_statistics(self, enabled: bool) -> Self

Enables or disables the collection of statistics after listing files

pub fn batch_size(&self) -> usize

Get the currently configured batch size

pub fn to_props(&self) -> HashMap<String, String>

Convert configuration options to name-value pairs with values converted to strings.

Note that this method will eventually be deprecated and replaced by config_options.

pub fn config_options(&self) -> &ConfigOptions

Return a handle to the configuration options.

pub fn config_options_mut(&mut self) -> &mut ConfigOptions

Return a mutable handle to the configuration options.

pub fn with_extension<T>(self, ext: Arc<T>) -> Selfwhere T: Send + Sync + 'static,

Add extensions.

Extensions can be used to attach extra data to the session config – e.g. tracing information or caches. Extensions are opaque and the types are unknown to DataFusion itself, which makes them extremely flexible. 1

Extensions are stored within an Arc so they do NOT require Clone. The are immutable. If you need to modify their state over their lifetime – e.g. for caches – you need to establish some for of interior mutability.

Extensions are indexed by their type T. If multiple values of the same type are provided, only the last one will be kept.

You may use get_extension to retrieve extensions.

Example

use std::sync::Arc;
use datafusion::execution::context::SessionConfig;

// application-specific extension types
struct Ext1(u8);
struct Ext2(u8);
struct Ext3(u8);

let ext1a = Arc::new(Ext1(10));
let ext1b = Arc::new(Ext1(11));
let ext2 = Arc::new(Ext2(2));

let cfg = SessionConfig::default()
    // will only remember the last Ext1
    .with_extension(Arc::clone(&ext1a))
    .with_extension(Arc::clone(&ext1b))
    .with_extension(Arc::clone(&ext2));

let ext1_received = cfg.get_extension::<Ext1>().unwrap();
assert!(!Arc::ptr_eq(&ext1_received, &ext1a));
assert!(Arc::ptr_eq(&ext1_received, &ext1b));

let ext2_received = cfg.get_extension::<Ext2>().unwrap();
assert!(Arc::ptr_eq(&ext2_received, &ext2));

assert!(cfg.get_extension::<Ext3>().is_none());

Compare that to ConfigOptions which only supports ScalarValue payloads. ↩

pub fn get_extension<T>(&self) -> Option<Arc<T>>where T: Send + Sync + 'static,

Get extension, if any for the specified type T exists.

See with_extension on how to add attach extensions.