Module datafusion::physical_optimizer::join_selection

Select the proper PartitionMode and build side based on the avaliable statistics for hash join.

Structs

JoinSelection

For hash join with the partition mode PartitionMode::Auto, JoinSelection rule will make a cost based decision to select which PartitionMode mode(Partitioned/CollectLeft) is optimal based on the available statistics that the inputs have. If the statistics information is not available, the partition mode will fall back to PartitionMode::Partitioned.

Functions

supports_swap

Predicate that checks whether the given join type supports input swapping.

swap_hash_join

This function swaps the inputs of the given join operator.

swap_join_type

This function returns the new join type we get after swapping the given join’s inputs.

swap_reverting_projection

When the order of the join is changed by the optimizer, the columns in the output should not be impacted. This function creates the expressions that will allow to swap back the values from the original left as the first columns and those on the right next.