Trait datafusion::physical_optimizer::optimizer::PhysicalOptimizerRule

pub trait PhysicalOptimizerRule {
    // Required methods
    fn optimize(
        &self,
        plan: Arc<dyn ExecutionPlan>,
        config: &ConfigOptions
    ) -> Result<Arc<dyn ExecutionPlan>>;
    fn name(&self) -> &str;
    fn schema_check(&self) -> bool;
}

PhysicalOptimizerRule transforms one [‘ExecutionPlan’] into another which computes the same results, but in a potentially more efficient way.

Required Methods

fn optimize( &self, plan: Arc<dyn ExecutionPlan>, config: &ConfigOptions ) -> Result<Arc<dyn ExecutionPlan>>

Rewrite plan to an optimized form

fn name(&self) -> &str

A human readable name for this optimizer rule

fn schema_check(&self) -> bool

A flag to indicate whether the physical planner should valid the rule will not change the schema of the plan after the rewriting. Some of the optimization rules might change the nullable properties of the schema and should disable the schema check.