Module datafusion::physical_plan::expressions

Defines physical expressions that can evaluated at runtime during query execution

Modules

helpers

Module with some convenient methods used in expression building

Structs

ApproxDistinct

APPROX_DISTINCT aggregate expression

ApproxMedian

MEDIAN aggregate expression

ApproxPercentileCont

APPROX_PERCENTILE_CONT aggregate expression

ApproxPercentileContWithWeight

APPROX_PERCENTILE_CONT_WITH_WEIGTH aggregate expression

ArrayAgg

ARRAY_AGG aggregate expression

Avg

AVG aggregate expression

AvgAccumulator

An accumulator to compute the average

BinaryExpr

Binary expression

BooleanQueryEvalFunc

BooleanQueryExpr

Binary expression

CaseExpr

The CASE expression is similar to a series of nested if/else and there are two forms that can be used. The first form consists of a series of boolean “when” expressions with corresponding “then” expressions, and an optional “else” expression.

CastExpr

CAST expression casts an expression to a specific data type and returns a runtime error on invalid cast

Column

Represents the column at a given index in a RecordBatch

Correlation

CORR aggregate expression

Count

COUNT aggregate expression Returns the amount of non-null values of the given expression.

Covariance

COVAR and COVAR_SAMP aggregate expression

CovariancePop

COVAR_POP aggregate expression

CumeDist

CumeDist calculates the cume_dist in the window function with order by

DateTimeIntervalExpr

Perform DATE/TIME/TIMESTAMP +/ INTERVAL math

DistinctArrayAgg

Expression for a ARRAY_AGG(DISTINCT) aggregation.

DistinctCount

Expression for a COUNT(DISTINCT) aggregation.

DistinctSum

Expression for a SUM(DISTINCT) aggregation.

Dnf

GetIndexedFieldExpr

expression to get a field of a struct array.

Grouping

GROUPING aggregate expression Returns the amount of non-null values of the given expression.

InListExpr

InList

IsNotNullExpr

IS NOT NULL expression

IsNullExpr

IS NULL expression

LikeExpr

Literal

Represents a literal value

Max

MAX aggregate expression

MaxAccumulator

An accumulator to compute the maximum value

Median

MEDIAN aggregate expression. This uses a lot of memory because all values need to be stored in memory before a result can be computed. If an approximation is sufficient then APPROX_MEDIAN provides a much more efficient solution.

Min

MIN aggregate expression

MinAccumulator

An accumulator to compute the minimum value

NegativeExpr

Negative expression

NoOp

A place holder expression, can not be evaluated.

NotExpr

Not expression

NthValue

nth_value expression

Ntile

PhysicalSortExpr

Represents Sort operation for a column in a RecordBatch

Rank

Rank calculates the rank in the window function with order by

RowNumber

row_number expression

Stddev

STDDEV and STDDEV_SAMP (standard deviation) aggregate expression

StddevPop

STDDEV_POP population aggregate expression

Sum

SUM aggregate expression

TryCastExpr

TRY_CAST expression casts an expression to a specific data type and retuns NULL on invalid cast

UnKnownColumn

Variance

VAR and VAR_SAMP aggregate expression

VariancePop

VAR_POP aggregate expression

WindowShift

window shift expression

Enums

RankType

StatsType

Enum used for differentiating population and sample for statistical functions

Constants

DEFAULT_DATAFUSION_CAST_OPTIONS

provide DataFusion default cast options

Functions

binary

Create a binary expression whose arguments are correctly coerced. This function errors if it is not possible to coerce the arguments to computational types supported by the operator.

boolean_query

boolean_query_with_cnf

case

Create a CASE expression

cast

Return a PhysicalExpression representing expr casted to cast_type, if any casting is needed.

cast_column

Internal cast function for casting ColumnarValue -> ColumnarValue for cast_type

cast_with_options

Return a PhysicalExpression representing expr casted to cast_type, if any casting is needed.

col

Create a column expression

create_aggregate_expr

Create a physical aggregation expression. This function errors when input_phy_exprs’ can’t be coerced to a valid argument type of the aggregation function.

cume_dist

Create a cume_dist window function

dense_rank

Create a dense rank window function

format_state_name

returns the name of the state

in_list

Creates a unary expression InList

is_not_null

Create an IS NOT NULL expression

is_null

Create an IS NULL expression

lag

lag() window function

lead

lead() window function

like

Create a like expression, erroring if the argument types are not compatible.

lit

Create a literal expression

negative

Creates a unary expression NEGATIVE

not

Creates a unary expression NOT

nullif_func

Implements NULLIF(expr1, expr2) Args: 0 - left expr is any array 1 - if the left is equal to this expr2, then the result is NULL, otherwise left value is passed.

percent_rank

Create a percent rank window function

rank

Create a rank window function

try_cast

Return a PhysicalExpression representing expr casted to cast_type, if any casting is needed.