Module datafusion::physical_plan::metrics

Metrics for recording information about execution

Structs

BaselineMetrics

Helper for creating and tracking common “baseline” metrics for each operator

CompositeMetricsSet

Collects all metrics during a complex operation, which is composed of multiple steps and each stage reports its statistics separately. Give sort as an example, when the dataset is more significant than available memory, it will report multiple in-mem sort metrics and final merge-sort metrics from SortPreservingMergeStream. Therefore, We need a separation of metrics for which are final metrics (for output_rows accumulation), and which are intermediate metrics that we only account for elapsed_compute time.

Count

A counter to record things such as number of input or output rows

ExecutionPlanMetricsSet

A set of Metrics for an individual “operator” (e.g. &dyn ExecutionPlan).

Gauge

A gauge is the simplest metrics type. It just returns a value. For example, you can easily expose current memory consumption with a gauge.

Label

name=value pairs identifiying a metric. This concept is called various things in various different systems:

MemTrackingMetrics

Wraps a BaselineMetrics and records memory usage on a MemoryReservation

Metric

Something that tracks a value of interest (metric) of a DataFusion ExecutionPlan execution.

MetricBuilder

Structure for constructing metrics, counters, timers, etc.

MetricsSet

A snapshot of the metrics for a particular (ExecutionPlan).

ScopedTimerGuard

RAAI structure that adds all time between its construction and destruction to the CPU time or the first call to stop whichever comes first

Time

Measure a potentially non contiguous duration of time

Timestamp

Stores a single timestamp, stored as the number of nanoseconds elapsed from Jan 1, 1970 UTC

Enums

MetricValue

Possible values for a super::Metric.

Traits

RecordOutput

Trait for things that produce output rows as a result of execution.