Struct datafusion::physical_plan::metrics::BaselineMetrics

pub struct BaselineMetrics { /* private fields */ }

Helper for creating and tracking common “baseline” metrics for each operator

Example:

use datafusion::physical_plan::metrics::{BaselineMetrics, ExecutionPlanMetricsSet};
let metrics = ExecutionPlanMetricsSet::new();

let partition = 2;
let baseline_metrics = BaselineMetrics::new(&metrics, partition);

// during execution, in CPU intensive operation:
let timer = baseline_metrics.elapsed_compute().timer();
// .. do CPU intensive work
timer.done();

// when operator is finished:
baseline_metrics.done();

Implementations

impl BaselineMetrics

pub fn new(metrics: &ExecutionPlanMetricsSet, partition: usize) -> Self

Create a new BaselineMetric structure, and set start_time to now

pub fn elapsed_compute(&self) -> &Time

return the metric for cpu time spend in this operator

pub fn spill_count(&self) -> &Count

return the metric for the total number of spills triggered during execution

pub fn spilled_bytes(&self) -> &Count

return the metric for the total spilled bytes during execution

pub fn mem_used(&self) -> &Gauge

return the metric for current memory usage

pub fn record_spill(&self, spilled_bytes: usize)

Record a spill of spilled_bytes size.

pub fn output_rows(&self) -> &Count

return the metric for the total number of output rows produced

pub fn done(&self)

Records the fact that this operator’s execution is complete (recording the end_time metric).

Note care should be taken to call done() manually if BaselineMetrics is not dropped immediately upon operator completion, as async streams may not be dropped immediately depending on the consumer.

pub fn record_output(&self, num_rows: usize)

Record that some number of rows have been produced as output

See the RecordOutput for conveniently recording record batch output for other thing

pub fn try_done(&self)

If not previously recorded done(), record

pub fn record_poll( &self, poll: Poll<Option<Result<RecordBatch>>> ) -> Poll<Option<Result<RecordBatch>>>

Process a poll result of a stream producing output for an operator, recording the output rows and stream done time and returning the same poll result