Trait datafusion::physical_plan::metrics::RecordOutput

pub trait RecordOutput {
    // Required method
    fn record_output(self, bm: &BaselineMetrics) -> Self;
}

Trait for things that produce output rows as a result of execution.

Required Methods

fn record_output(self, bm: &BaselineMetrics) -> Self

Record that some number of output rows have been produced

Meant to be composable so that instead of returning batch the operator can return batch.record_output(baseline_metrics)