Module datafusion::physical_plan::planner

Physical query planner

Structs

DefaultPhysicalPlanner

Default single node physical query planner that converts a LogicalPlan to an ExecutionPlan suitable for execution.

Traits

ExtensionPlanner

This trait exposes the ability to plan an ExecutionPlan out of a LogicalPlan.

PhysicalPlanner

Physical query planner that converts a LogicalPlan to an ExecutionPlan suitable for execution.

Functions

create_aggregate_expr

Create an aggregate expression from a logical expression or an alias

create_aggregate_expr_with_name

Create an aggregate expression with a name from a logical expression

create_physical_sort_expr

Create a physical sort expression from a logical expression

create_window_expr

Create a window expression from a logical expression or an alias

create_window_expr_with_name

Create a window expression with a name from a logical expression

is_window_valid

Check if window bounds are valid after schema information is available, and window_frame bounds are casted to the corresponding column type. queries like: OVER (ORDER BY a RANGES BETWEEN 3 PRECEDING AND 5 PRECEDING) OVER (ORDER BY a RANGES BETWEEN INTERVAL ‘3 DAY’ PRECEDING AND ‘5 DAY’ PRECEDING) are rejected