Trait datafusion::physical_plan::RecordBatchStream

pub trait RecordBatchStream: Stream<Item = Result<RecordBatch>> {
    // Required method
    fn schema(&self) -> SchemaRef;
}

Trait for types that stream arrow::record_batch::RecordBatch

Required Methods

fn schema(&self) -> SchemaRef

Returns the schema of this RecordBatchStream.

Implementation of this trait should guarantee that all RecordBatch’s returned by this stream should have the same schema as returned from this method.