Trait datafusion::physical_plan::WindowExpr

pub trait WindowExpr: Send + Sync + Debug {

    // Required methods
    fn as_any(&self) -> &(dyn Any + 'static);
    fn field(&self) -> Result<Field, DataFusionError>;
    fn expressions(&self) -> Vec<Arc<dyn PhysicalExpr + 'static>, Global> ;
    fn evaluate(
        &self,
        batch: &RecordBatch
    ) -> Result<Arc<dyn Array + 'static>, DataFusionError>;
    fn partition_by(&self) -> &[Arc<dyn PhysicalExpr + 'static>];
    fn order_by(&self) -> &[PhysicalSortExpr];
    fn get_window_frame(&self) -> &Arc<WindowFrame>;
    fn uses_bounded_memory(&self) -> bool;
    fn get_reverse_expr(&self) -> Option<Arc<dyn WindowExpr + 'static>>;

    // Provided methods
    fn name(&self) -> &str { ... }
    fn evaluate_args(
        &self,
        batch: &RecordBatch
    ) -> Result<Vec<Arc<dyn Array + 'static>, Global>, DataFusionError> { ... }
    fn evaluate_stateful(
        &self,
        _partition_batches: &IndexMap<Vec<ScalarValue, Global>, PartitionBatchState, RandomState>,
        _window_agg_state: &mut IndexMap<Vec<ScalarValue, Global>, WindowState, RandomState>
    ) -> Result<(), DataFusionError> { ... }
    fn evaluate_partition_points(
        &self,
        num_rows: usize,
        partition_columns: &[SortColumn]
    ) -> Result<Vec<Range<usize>, Global>, DataFusionError> { ... }
    fn order_by_columns(
        &self,
        batch: &RecordBatch
    ) -> Result<Vec<SortColumn, Global>, DataFusionError> { ... }
    fn sort_columns(
        &self,
        batch: &RecordBatch
    ) -> Result<Vec<SortColumn, Global>, DataFusionError> { ... }
    fn get_values_orderbys(
        &self,
        record_batch: &RecordBatch
    ) -> Result<(Vec<Arc<dyn Array + 'static>, Global>, Vec<Arc<dyn Array + 'static>, Global>), DataFusionError> { ... }

}

A window expression that:

knows its resulting field

Required Methods

fn as_any(&self) -> &(dyn Any + 'static)

Returns the window expression as Any so that it can be downcast to a specific implementation.

fn field(&self) -> Result<Field, DataFusionError>

The field of the final result of this window function.

fn expressions(&self) -> Vec<Arc<dyn PhysicalExpr + 'static>, Global>

Expressions that are passed to the WindowAccumulator. Functions which take a single input argument, such as sum, return a single datafusion_expr::expr::Expr, others (e.g. cov) return many.

fn evaluate( &self, batch: &RecordBatch ) -> Result<Arc<dyn Array + 'static>, DataFusionError>

Evaluate the window function values against the batch

fn partition_by(&self) -> &[Arc<dyn PhysicalExpr + 'static>]

Expressions that’s from the window function’s partition by clause, empty if absent

fn order_by(&self) -> &[PhysicalSortExpr]

Expressions that’s from the window function’s order by clause, empty if absent

fn get_window_frame(&self) -> &Arc<WindowFrame>

Get the window frame of this WindowExpr.

fn uses_bounded_memory(&self) -> bool

Return a flag indicating whether this WindowExpr can run with bounded memory.

fn get_reverse_expr(&self) -> Option<Arc<dyn WindowExpr + 'static>>

Get the reverse expression of this WindowExpr.

Provided Methods

fn name(&self) -> &str

Human readable name such as "MIN(c2)" or "RANK()". The default implementation returns placeholder text.

fn evaluate_args( &self, batch: &RecordBatch ) -> Result<Vec<Arc<dyn Array + 'static>, Global>, DataFusionError>

Evaluate the window function arguments against the batch and return array ref, normally the resulting Vec is a single element one.

fn evaluate_stateful( &self, _partition_batches: &IndexMap<Vec<ScalarValue, Global>, PartitionBatchState, RandomState>, _window_agg_state: &mut IndexMap<Vec<ScalarValue, Global>, WindowState, RandomState> ) -> Result<(), DataFusionError>

Evaluate the window function against the batch. This function facilitates stateful, bounded-memory implementations.

fn evaluate_partition_points( &self, num_rows: usize, partition_columns: &[SortColumn] ) -> Result<Vec<Range<usize>, Global>, DataFusionError>

Evaluate the partition points given the sort columns; if the sort columns are empty then the result will be a single element Vec of the whole column rows.

fn order_by_columns( &self, batch: &RecordBatch ) -> Result<Vec<SortColumn, Global>, DataFusionError>

Get order by columns, empty if absent

fn sort_columns( &self, batch: &RecordBatch ) -> Result<Vec<SortColumn, Global>, DataFusionError>

Get sort columns that can be used for peer evaluation, empty if absent

fn get_values_orderbys( &self, record_batch: &RecordBatch ) -> Result<(Vec<Arc<dyn Array + 'static>, Global>, Vec<Arc<dyn Array + 'static>, Global>), DataFusionError>

Get values columns (argument of Window Function) and order by columns (columns of the ORDER BY expression) used in evaluators