Struct datafusion::physical_plan::union::UnionExec

pub struct UnionExec { /* private fields */ }

UnionExec: UNION ALL execution plan.

UnionExec combines multiple inputs with the same schema by concatenating the partitions. It does not mix or copy data within or across partitions. Thus if the input partitions are sorted, the output partitions of the union are also sorted.

For example, given a UnionExec of two inputs, with N partitions, and M partitions, there will be N+M output partitions. The first N output partitions are from Input 1 partitions, and then next M output partitions are from Input 2.

                      ▲       ▲           ▲         ▲
                      │       │           │         │
    Output            │  ...  │           │         │
  Partitions          │0      │N-1        │ N       │N+M-1
(passes through   ┌────┴───────┴───────────┴─────────┴───┐
the N+M input    │              UnionExec               │
 partitions)     │                                      │
                 └──────────────────────────────────────┘
                                     ▲
                                     │
                                     │
      Input           ┌────────┬─────┴────┬──────────┐
    Partitions        │ ...    │          │     ...  │
                   0  │        │ N-1      │ 0        │  M-1
                 ┌────┴────────┴───┐  ┌───┴──────────┴───┐
                 │                 │  │                  │
                 │                 │  │                  │
                 │                 │  │                  │
                 │                 │  │                  │
                 │                 │  │                  │
                 │                 │  │                  │
                 │Input 1          │  │Input 2           │
                 └─────────────────┘  └──────────────────┘

Implementations

impl UnionExec

pub fn try_new_with_schema( inputs: Vec<Arc<dyn ExecutionPlan>>, schema: DFSchemaRef ) -> Result<Self>

Create a new UnionExec with specified schema. The schema should always be a subset of the schema of inputs, otherwise, an error will be returned.

pub fn new(inputs: Vec<Arc<dyn ExecutionPlan>>) -> Self

Create a new UnionExec

pub fn inputs(&self) -> &Vec<Arc<dyn ExecutionPlan>>

Get inputs of the execution plan